Re-worked the movement and plowing
Today I reworked the plowing mechaninc and the player script. I added a base class from which the player now inherits from so that that we can expand and make more characters with the base functionality of the character class. I also changed the way the player can plow the tiles. Before it was solely based on if you are on top of the tile. Now I used a 2D array to keep track of all the tiles in the level. Then used an OnTriggerEnter to keep track of which tile the player is currently standing on. Then depending on which direction the player is facing, I set the selected tile to be the one directly in front of the player. This gives us accuracy on where the player is going to plow based on player input.  I also switched the art assets out that way we have a decent looking prototype while we wait for art. I also removed the Farming RPG package because I am making custom code to beter suit the project goals